@@ -41,10 +41,10 @@ class App extends StatelessWidget {
           // outlinedButtonTheme: OutlinedButtonThemeData(style: OutlinedButton.styleFrom(primary: Style.pinkDark)),
           // textButtonTheme: TextButtonThemeData(style: TextButton.styleFrom(primary: Style.pinkDark)),
           textTheme: TextTheme(
-            headline5: Style.titleTextStyle,
-            headline6: Style.headerTextStyle,
-            bodyText2: Style.regularTextStyle,
-            bodyText1: Style.regularTextStyle,
+            headlineSmall: Style.titleTextStyle,
+            titleLarge: Style.headerTextStyle,
+            bodyMedium: Style.regularTextStyle,
+            bodyLarge: Style.regularTextStyle,
           ),
         ),
         debugShowCheckedModeBanner: false,
