@@ -1,5 +1,5 @@
 import 'package:flutter/foundation.dart';
-import 'package:package_info/package_info.dart';
+import 'package:package_info_plus/package_info_plus.dart';
 
 class InfoProviderModel extends ChangeNotifier {
   bool _initialized = false;
